Add files via upload
This is the entire DQA application directory structure.  Just download and extract.  You can use this to upload to Google Drive and then use on Google Colab to view the Jupyter notebooks.  Just use the upload feature of Google Drive and add it to the root of My Drive.  The Google Drive path should look like:  My Drive/DQA.  That's it!